+++ src/main/java/gregtech/api/modernmaterials/blocks/blocktypes/blockof/BlockOfSimpleBlockRenderer.java
@@ -1,6 +1,6 @@
 package gregtech.api.modernmaterials.blocks.blocktypes.blockof;
 
-import static gregtech.api.modernmaterials.render.Utilities.drawBlock;
+import static gregtech.api.modernmaterials.render.ModernMaterialRenderUtil.drawBlock;
 
 import java.awt.Color;
