+++ src/main/java/gregtech/api/modernmaterials/blocks/blocktypes/blockof/special/UniversiumBlockOfBlockRenderer.java
@@ -1,6 +1,6 @@
 package gregtech.api.modernmaterials.blocks.blocktypes.blockof.special;
 
-import static gregtech.api.modernmaterials.render.Utilities.renderBlock;
+import static gregtech.api.modernmaterials.render.ModernMaterialRenderUtil.renderBlock;
 
 import net.minecraft.client.renderer.tileentity.TileEntitySpecialRenderer;
 import net.minecraft.tileentity.TileEntity;
