+++ src/main/java/gregtech/api/modernmaterials/blocks/blocktypes/framebox/special/UniversiumFrameBlockRenderer.java
@@ -1,6 +1,6 @@
 package gregtech.api.modernmaterials.blocks.blocktypes.framebox.special;
 
-import static gregtech.api.modernmaterials.render.Utilities.renderBlock;
+import static gregtech.api.modernmaterials.render.ModernMaterialRenderUtil.renderBlock;
 
 import net.minecraft.client.renderer.tileentity.TileEntitySpecialRenderer;
 import net.minecraft.tileentity.TileEntity;
