+++ src/main/java/gregtech/api/modernmaterials/blocks/blocktypes/ore/OreBlockRenderer.java
@@ -1,6 +1,6 @@
 package gregtech.api.modernmaterials.blocks.blocktypes.ore;
 
-import static gregtech.api.modernmaterials.render.Utilities.drawBlock;
+import static gregtech.api.modernmaterials.render.ModernMaterialRenderUtil.drawBlock;
 
 import java.awt.*;
